requirements.txt
pandas>=2.0,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.5,<2.0
openpyxl>=3.1,<4.0
psycopg2-binary>=2.9,<3.0
cryptography>=42.0,<46.0
django-filter>=24.0,<26.0
django-cors-headers>=4.4,<5.0
